fix(job): component separation job detail feat: share job
